backend/foodgram/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pillow==9.5.0
django-filter==23.2
djoser==2.2.0
python_dotenv==1.0.0